# docs/openapi/public.yaml
openapi: 3.1.0
info:
  title: City Flood MVP API
  version: 0.1.0
paths:
...
  /v1/cities/{city_id}/status:
    get:
      tags:
      - analytics
      summary: City Status
      operationId: city_status_v1_cities__city_id__status_get
      parameters:
      - name: city_id
        in: path
        required: true
        schema:
          type: string
          title: City Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityStatusOut'
              example:
                city_id: city_porto_mvp
                risk: low
                active_events: 0
                hotspots: 0
                updated_at: '2026-02-02T12:23:13Z'
                request_id: req_example
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/cities/{city_id}/summary:
    get:
      tags:
      - analytics
      summary: City Summary
      operationId: city_summary_v1_cities__city_id__summary_get
      parameters:
      - name: city_id
        in: path
        required: true
        schema:
          type: string
          title: City Id
      - name: minutes
        in: query
        required: false
        schema:
          type: integer
          maximum: 1440
          minimum: 1
          default: 15
          title: Minutes
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitySummaryOut'
              example:
                city_id: city_porto_mvp
                now: '2026-02-02T12:00:00Z'
                rain_mmph: 3.2
                river_level_m: 1.8
                status_counts:
                  normal: 120
                  watch: 5
                  warning: 1
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/hotspots:
    get:
      tags:
      - analytics
      summary: Hotspots
      operationId: hotspots_v1_hotspots_get
      parameters:
      - name: city_id
        in: query
        required: true
        schema:
          type: string
          title: City Id
      - name: metric
        in: query
        required: false
        schema:
          type: string
          default: overflow_risk
          title: Metric
      - name: top
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Top
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HotspotOut'
                title: Response Hotspots V1 Hotspots Get
              example:
              - asset_id: pipe_123
                score: 42.5
                details: {}
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/analytics/runs:
    get:
      tags:
      - analytics
      summary: Analytics Runs
      operationId: analytics_runs_v1_analytics_runs_get
      parameters:
      - name: city_id
        in: query
        required: true
        schema:
          type: string
          title: City Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnalyticsRunOut'
                title: Response Analytics Runs V1 Analytics Runs Get
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/analytics/jobs:
    post:
      tags:
      - analytics
      summary: Create Job
      operationId: create_job_v1_analytics_jobs_post
      parameters:
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportJobRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJobOut'
        '401':
          $ref: '#/components/responses/UnauthorizedApiKey'
        '403':
          $ref: '#/components/responses/ForbiddenApiKey'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
      security:
      - ApiKeyAuth: []
  /v1/analytics/jobs/{job_id}/download:
    get:
      tags:
      - analytics
      summary: Download Job
      operationId: download_job_v1_analytics_jobs__job_id__download_get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '401':
          $ref: '#/components/responses/UnauthorizedApiKey'
        '403':
          $ref: '#/components/responses/ForbiddenApiKey'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
      security:
      - ApiKeyAuth: []
...
components:
  schemas:
    AnalyticsRunOut:
      properties:
        run_id:
          type: string
          title: Run Id
        city_id:
          type: string
          title: City Id
        start_ts:
          type: string
          format: date-time
          title: Start Ts
        end_ts:
          type: string
          format: date-time
          title: End Ts
        status:
          type: string
          title: Status
        version:
          type: string
          title: Version
        params:
          additionalProperties: true
          type: object
          title: Params
        metrics:
          additionalProperties: true
          type: object
          title: Metrics
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - run_id
      - city_id
      - start_ts
      - end_ts
      - status
      - version
      - params
      - metrics
      - created_at
      - updated_at
      title: AnalyticsRunOut
...
    CityStatusOut:
      properties:
        city_id:
          type: string
          title: City Id
        risk:
          type: string
          title: Risk
        active_events:
          type: integer
          title: Active Events
        hotspots:
          type: integer
          title: Hotspots
        updated_at:
          type: string
          format: date-time
          title: Updated At
        request_id:
          type: string
          title: Request Id
      type: object
      required:
      - city_id
      - risk
      - active_events
      - hotspots
      - updated_at
      - request_id
      title: CityStatusOut
    CitySummaryOut:
      properties:
        city_id:
          type: string
          title: City Id
        now:
          type: string
          format: date-time
          title: Now
        rain_mmph:
          type: number
          title: Rain Mmph
        river_level_m:
          type: number
          title: River Level M
        status_counts:
          additionalProperties: true
          type: object
          title: Status Counts
      type: object
      required:
      - city_id
      - now
      - rain_mmph
      - river_level_m
      - status_counts
      title: CitySummaryOut
...
    ExportJobOut:
      properties:
        job_id:
          type: string
          title: Job Id
        job_type:
          type: string
          title: Job Type
        status:
          type: string
          title: Status
        progress:
          type: number
          title: Progress
        file_path:
          anyOf:
          - type: string
          - type: 'null'
          title: File Path
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        row_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Row Count
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - job_id
      - job_type
      - status
      - progress
      - created_at
      - updated_at
      title: ExportJobOut
    ExportJobQuery:
      properties:
        asset_ids:
          items:
            type: string
          type: array
          title: Asset Ids
        metric:
          type: string
          title: Metric
        from:
          type: string
          format: date-time
          title: From
        to:
          type: string
          format: date-time
          title: To
        granularity:
          type: string
          title: Granularity
          default: 5m
        agg:
          type: string
          title: Agg
          default: avg
      type: object
      required:
      - asset_ids
      - metric
      - from
      - to
      title: ExportJobQuery
    ExportJobRequest:
      properties:
        type:
          type: string
          const: export_csv
          title: Type
        query:
          $ref: '#/components/schemas/ExportJobQuery'
      type: object
      required:
      - type
      - query
      title: ExportJobRequest
...
    HotspotOut:
      properties:
        asset_id:
          type: string
          title: Asset Id
        score:
          type: number
          title: Score
        details:
          additionalProperties: true
          type: object
          title: Details
      type: object
      required:
      - asset_id
      - score
      title: HotspotOut
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            request_id:
              type: string
          required:
          - code
          - message
          - details
          - request_id
      required:
      - error
  responses:
    InvalidArgument:
      description: 'INVALID_ARGUMENT: Validation error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INVALID_ARGUMENT
              message: Validation error
              details: {}
              request_id: req_example
    UnauthorizedApiKey:
      description: 'UNAUTHORIZED: Missing API key'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: Missing API key
              details: {}
              request_id: req_example
    ForbiddenApiKey:
      description: 'FORBIDDEN: Invalid API key'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FORBIDDEN
              message: Invalid API key
              details: {}
              request_id: req_example
...
    NotFound:
      description: 'NOT_FOUND: Resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
              details: {}
              request_id: req_example
...
    UnprocessableEntity:
      description: 'UNPROCESSABLE_ENTITY: Unprocessable entity'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNPROCESSABLE_ENTITY
              message: Unprocessable entity
              details: {}
              request_id: req_example
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security: []